ENH: Refactor source code to use fmu-datamodels
