+++ src/fmu/dataio/providers/objectdata/_faultroom.py
@@ -7,13 +7,13 @@ from typing import TYPE_CHECKING, Final
 
 from fmu.dataio._definitions import ExportFolder, FileExtension
 from fmu.dataio._logging import null_logger
-from fmu.dataio._models.fmu_results.data import BoundingBox3D
-from fmu.dataio._models.fmu_results.enums import FileFormat, Layout, ObjectMetadataClass
-from fmu.dataio._models.fmu_results.global_configuration import (
+from fmu.dataio.providers.objectdata._utils import Utils
+from fmu.datamodels.fmu_results.data import BoundingBox3D
+from fmu.datamodels.fmu_results.enums import FileFormat, Layout, ObjectMetadataClass
+from fmu.datamodels.fmu_results.global_configuration import (
     GlobalConfiguration,
 )
-from fmu.dataio._models.fmu_results.specification import FaultRoomSurfaceSpecification
-from fmu.dataio.providers.objectdata._utils import Utils
+from fmu.datamodels.fmu_results.specification import FaultRoomSurfaceSpecification
 
 from ._base import (
     ObjectDataProvider,
