+++ src/fmu/dataio/providers/objectdata/_provider.py
@@ -97,9 +97,9 @@ import xtgeo
 
 from fmu.dataio._definitions import ExportFolder, FileExtension
 from fmu.dataio._logging import null_logger
-from fmu.dataio._models.fmu_results.enums import FileFormat, Layout, ObjectMetadataClass
 from fmu.dataio._readers.faultroom import FaultRoomSurface
 from fmu.dataio._readers.tsurf import TSurfData
+from fmu.datamodels.fmu_results.enums import FileFormat, Layout, ObjectMetadataClass
 
 from ._base import (
     ObjectDataProvider,
@@ -119,9 +119,9 @@ from ._xtgeo import (
 if TYPE_CHECKING:
     from io import BytesIO
 
-    from fmu.dataio._models.fmu_results.standard_result import StandardResult
     from fmu.dataio.dataio import ExportData
     from fmu.dataio.types import Inferrable
+    from fmu.datamodels.fmu_results.standard_result import StandardResult
 
 logger: Final = null_logger(__name__)
