+++ src/fmu/dataio/providers/objectdata/_triangulated_surface.py
@@ -6,12 +6,12 @@ from typing import TYPE_CHECKING, Final
 
 from fmu.dataio._definitions import ExportFolder, FileExtension
 from fmu.dataio._logging import null_logger
-from fmu.dataio._models.fmu_results.data import BoundingBox3D
-from fmu.dataio._models.fmu_results.enums import FileFormat, Layout, ObjectMetadataClass
-from fmu.dataio._models.fmu_results.specification import (
+from fmu.dataio._readers import tsurf as reader
+from fmu.datamodels.fmu_results.data import BoundingBox3D
+from fmu.datamodels.fmu_results.enums import FileFormat, Layout, ObjectMetadataClass
+from fmu.datamodels.fmu_results.specification import (
     TriangulatedSurfaceSpecification,
 )
-from fmu.dataio._readers import tsurf as reader
 
 from ._base import (
     ObjectDataProvider,
